Moved the wave colour scheme to the new setting system
Fixed some bugs with the property table and fractal settings frame

+++ fractals/MandelBrotFractal.java
@@ -14,7 +14,7 @@ import util.math.Scale;
 public class MandelBrotFractal extends AbstractFractal {
 
 	private int maxIterations;
-	private double escapeValue;
+	private float escapeValue;
 
 	public MandelBrotFractal() {
 
@@ -24,7 +24,7 @@ public class MandelBrotFractal extends AbstractFractal {
 		// Set default values
 		name = "Mandelbrot Set";
 		maxIterations = 512;
-		escapeValue = 2.0;
+		escapeValue = 2.0f;
 		scale = new Scale(new Point(-0.5, 0));
 
 		// Create color scheme manager
@@ -60,13 +60,13 @@ public class MandelBrotFractal extends AbstractFractal {
 
 	}
 
-	public double getEscapeValue() {
+	public float getEscapeValue() {
 
 		return escapeValue;
 
 	}
 
-	public void setEscapeValue(double escapeValue) {
+	public void setEscapeValue(float escapeValue) {
 
 		this.escapeValue = escapeValue;
